+++ CHANGELOG.rst
@@ -15,6 +15,11 @@ latest
     levels; *default solver verbosity is new 1*; ``log`` can now be overwritten
     in ``solver_opts`` (mainly for debugging).
 
+- There is no ``optimize`` module any longer. The ``misfit`` and ``gradient``
+  are directly included in the Simulation class. (This is backwards
+  incompatible, but *optimize* was not yet declared stable; it is getting there
+  soon though.)
+
 - Bug fixes:
 
   - Track order when saving to hdf5.
