--- docs/api/optimize.rst
@@ -1,6 +1,0 @@
-Optimize
-========
-
-.. automodapi:: emg3d.optimize
-   :no-inheritance-diagram:
-   :no-heading:
